frontend

+++ frontend/js/cart.js
@@ -1,8 +1,10 @@
 var products = [],
 localStorageProducts = JSON.parse(localStorage.getItem("products"));
 //console.log(localStorageProducts)
-const items = document.getElementById("cart__items");
-const section = document.getElementById("cart__items");
+
+const items = document.getElementById("cart__items"),
+orderButton = document.getElementById("order"),
+form = document.querySelector('.cart__order__form');
 
 /*
  * Envoie une requet ver le serveur pour récupérer les informations sur les produits .
@@ -13,12 +15,16 @@ if ( localStorageProducts && localStorageProducts.length != 0 ) {
         .then(data => {
             // console.log(data);
             products = data;
-
+            
+            // mets en ordre le localStorage
             const productsListOrdered = localStorageProducts.sort((a, b) => {
                 if (a.nameItem < b.nameItem) {return -1;}
                 if (a.nameItem > b.nameItem) {return 1};
                 return 0;
             });
+            // console.log(productsListOrdered);
+
+            // créer chaque article de productsListOrdered avec les informations récupérer du coté de l'API.
             for (const item of productsListOrdered) {
                 const {idItem, colorItem, qtity} = item;
                 createItem({idItem, colorItem, qtity})
@@ -26,23 +32,9 @@ if ( localStorageProducts && localStorageProducts.length != 0 ) {
             updateQuantityAndPrice();
             
         }))
-        .catch(err => console.log(err));
+        .catch(err => error(err));
 } else {
-    // creation de message
-    const message = `
-        <p style="text-align: center; margin-bottom: 50px;" >
-            Votre panier est vide pour le moment
-        </p>
-    `
-    //imbrication de message dans la section cart__items
-    items.innerHTML += message;
-
-    //ne pas afficher les blocs .cart__price et .cart__order
-    const cartPrice = document.querySelector('.cart__price'),
-    cartOrder = document.querySelector('.cart__order');
-
-    cartPrice.style.display='none';
-    cartOrder.style.display='none';
+    error();
 }
 
 
@@ -54,52 +46,56 @@ if ( localStorageProducts && localStorageProducts.length != 0 ) {
  * @param { String } qtity
  * 
  */
-function createItem({idItem, colorItem, qtity}) {
+createItem = ({idItem, colorItem, qtity}) => {
     index = products.findIndex( item => item._id  == idItem );
-    const { name, price, imageUrl, description, altTxt } = products[index];
 
-    // créer l'article
-    const article = document.createElement("article");
-    article.innerHTML = `
-        <article class="cart__item" data-id="${idItem}" data-color="${colorItem}">
-            <div class="cart__item__img">
-                <img src="${imageUrl}" alt="${altTxt}">
-            </div>
-            <div class="cart__item__content">
-                <div class="cart__item__content__description">
-                    <h2>${name}</h2>
-                    <p>${colorItem}</p>
-                    <p>${price} €</p>
+    // Si l'article existe dans products alors on le créer
+    if (index != -1) {
+        const { name, price, imageUrl, description, altTxt } = products[index];
+
+        // créer l'article
+        const article = document.createElement("article");
+        article.innerHTML = `
+            <article class="cart__item" data-id="${idItem}" data-color="${colorItem}">
+                <div class="cart__item__img">
+                    <img src="${imageUrl}" alt="${altTxt}">
                 </div>
-                <div class="cart__item__content__settings">
-                    <div class="cart__item__content__settings__quantity">
-                        <p>Qté :</p>
-                        <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${qtity}">
+                <div class="cart__item__content">
+                    <div class="cart__item__content__description">
+                        <h2>${name}</h2>
+                        <p>${colorItem}</p>
+                        <p>${price} €</p>
                     </div>
-                    <div class="cart__item__content__settings__delete">
-                        <p class="deleteItem">Supprimer</p>
+                    <div class="cart__item__content__settings">
+                        <div class="cart__item__content__settings__quantity">
+                            <p>Qté :</p>
+                            <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${qtity}">
+                        </div>
+                        <div class="cart__item__content__settings__delete">
+                            <p class="deleteItem">Supprimer</p>
+                        </div>
                     </div>
                 </div>
-            </div>
-        </article>
-    `;
+            </article>
+        `;
 
-    // si la quantité de l'article change 
-    const itemQuantity = article.querySelector('.itemQuantity');
-    itemQuantity.addEventListener('change', onChangeQtity  );
-    
-    // si on clique sur supprimer
-    const deleteItem = article.querySelector('.deleteItem');
-    deleteItem.addEventListener('click', removeItem);
+        // si la quantité de l'article change 
+        const itemQuantity = article.querySelector('.itemQuantity');
+        itemQuantity.addEventListener('change', onChangeQtity  );
+        
+        // si on clique sur supprimer
+        const deleteItem = article.querySelector('.deleteItem');
+        deleteItem.addEventListener('click', removeItem);
 
-    // Imbrication de l'article dans la section
-    items.appendChild(article);
+        // Imbrication de l'article dans la section
+        items.appendChild(article);
+    }
 }
 
 /**
  * Mets à jour la quantité ainsi que le prix total
  */
-function updateQuantityAndPrice() { 
+updateQuantityAndPrice = () => { 
     let 
         totalQuantity = 0,
         totalPrice = 0;
@@ -139,7 +135,12 @@ function updateQuantityAndPrice() {
         }
     }
 
-function removeItem (e){
+/**
+ * si on clique sur le boutton supprimer celà éfface l'article
+ * 
+ * @param {any} e 
+ */
+removeItem = (e) => {
     const {id, color} = e.target.closest("article").dataset,
     productIndex = localStorageProducts.findIndex( item => item.idItem + item.colorItem == id+color );
 
@@ -150,19 +151,19 @@ function removeItem (e){
 
 
 
-// order
-const 
-    orderButton = document.getElementById("order"),
-    form = document.querySelector('.cart__order__form');
-
+// Pour chaque changement input dans le formulaire validate vérifie s'il y a des erreurs
 for (let i = 0; i < form.children.length-1; i++) {
     const input = form.children[i];
-    input.onchange = (e) => {
-        validate(e);
-    }
-    
+    input.onchange = e => validate(e);
 }
-const errorMessage = (name) => {
+
+/**
+ * Cette fonction retourne le message d'erreur en function du nom de chaque input du formulaire
+ * 
+ * @param { String } name 
+ * @returns { String }
+ */
+errorMessage = (name) => {
     if (name == 'firstName') {
         return "Veuillez saisir votre prénom !";
     } else if (name == 'lastName') {
@@ -174,34 +175,49 @@ const errorMessage = (name) => {
     } else if (name == 'email') {
         return "Veuillez saisir votre adresse email !";
     } else {
-        return "error"
+        return "erreur"
     }
 }
 
-const regex = (name) => {
-    if(name == 'firstName' || name =='lastName' || name == 'city' ) {
-        return /^[^@&"()!_$*€£`+=\/;?#\d]+$/;
-    } else if(name == 'address') {
-        return /(?!^\d+$)^[^@&"()!_$*€£`+=\/;?#]+$/;
+/**
+ * Regex est utilisé pour étudier les correspondances de texte (value) avec un motif (pattern) donné.
+ * 
+ * @param {String} value
+ * @param {String} name 
+ * @returns {Boolean}
+ */
+regex = (name, value) => {
+    let pattern = /^[^@&"()!_$*€£`+=\/;?#\d]+$/;
+    if(name == 'address') {
+        pattern = /(?!^\d+$)^[^@&"()!_$*€£`+=\/;?#]+$/;
     } else if (name == 'email') {
-        return /^[a-zA-Z0-9.]+@[a-zA-Z]+\.[a-zA-Z]{2,}$/i;
+        pattern = /^[a-zA-Z0-9.]+@[a-zA-Z]+\.[a-zA-Z]{2,}$/i;
     }
+    return pattern.test(value);
 }
-function validate(e) {
+
+/**
+ * Valider l'input, s'il y a une erreur l'afficher
+ * 
+ * @param {any} e 
+ */
+validate = (e) => {
     const { name, value, style } = e.target,
     errorMsg = document.getElementById(name+"ErrorMsg");
 
-    if (regex(name).test(value) == true ) {
+    if (regex(name, value)) {
         style.border = "0";
         errorMsg.innerText = "";
-        return 0;
     } else {
         style.border = "2px solid red"
         errorMsg.innerText = errorMessage(name);
-        return 1;
     }
 }
-orderButton.addEventListener('click', ()=>{
+
+// Lorsque on clique sur commander
+orderButton.addEventListener('click', () => {
+
+    //verifie chaque input dans le formulaire et afficher l'erreur s'il y'en a
     let error = 0;
     for (let i = 0; i < form.length-1; i++) {
         const e = form[i],
@@ -212,7 +228,8 @@ orderButton.addEventListener('click', ()=>{
             error += 1;
         }
     }
-        //alert("Veuillez remplir tous les champs")
+    
+    // s'il n'y a pas d'erreur alors on envoie la requette à l'API
     if (error == 0) {
         const formOrder = {
             firstName: firstName.value,
@@ -226,19 +243,24 @@ orderButton.addEventListener('click', ()=>{
     
 })
 
-
-function requestOrder(formOrder) {
-    // Données à envoyer
+/**
+ * fait la demmande de commande à l'API
+ * 
+ * @param {Object} formOrder
+ */
+requestOrder = (formOrder) => {
     const localStorageProductsID = []
     for (const item of localStorageProducts) {
         localStorageProductsID.push(item.idItem)
     }
 
+    // Données à envoyer
     const order = {
         contact: formOrder,
         products: localStorageProductsID
     };
-    console.log(order)
+    // test n°  | console.log(order)
+
     // En-tête de la requête
     const entete = {
         method: "POST",
@@ -248,7 +270,10 @@ function requestOrder(formOrder) {
 
     fetch("http://localhost:3000/api/products/order", entete)
         .then((response) => {
-            // On vérifie qu'on reçoit bien un status HTTP 201 c'est à dire que la requête a réussi et qu'une ressource a été créée en conséquence
+            /**
+             * On vérifie qu'on reçoit bien un status HTTP 201
+             * c'est à dire que la requête a réussi et qu'une ressource a été créée en conséquence
+             */
             if (response.status == 201) {
                 return response.json();
             }
@@ -266,4 +291,24 @@ function requestOrder(formOrder) {
               console.log(error);
             }
         );
+}
+
+function error(err) {
+    //test n° | console.log(err)
+
+    // creation de message
+    const message = `
+        <p style="text-align: center; margin-bottom: 50px;" >
+            Votre panier est vide pour le moment
+        </p>
+    `
+    //imbrication de message dans la section cart__items
+    items.innerHTML += message;
+
+    //ne pas afficher les blocs .cart__price et .cart__order
+    const cartPrice = document.querySelector('.cart__price'),
+    cartOrder = document.querySelector('.cart__order');
+
+    cartPrice.style.display='none';
+    cartOrder.style.display='none';
 }
